First draft
I have been unable to prove the Hom-Tensor adjunction (Theorem 1.4 in this version) in the category of base-point topological spaces. I could construct the bijection but could not prove continuity in either direction. 
In the CW-decomposition of the classical surfaces, I could not prove that any of the constructions actually correspond to the other descriptions of these surfaces. Nor could I prove that these constructions require the least number of cells. 
I have not typed out the problem on Schubert decomposition of the projective spaces.